+++ R/regions.R
@@ -6,8 +6,10 @@
 #'
 #' @details Expecting total of 4 bounding box arguments, If arguments are unnamed they should be in the following order: "xmin", "ymax", "xmax", "ymin".
 #'
-#' @param ... Either a single terra `SpatExtent` object or arguments: `xmin`/`ymax`/`xmax`/`ymin`. If the four bounding arguments are not named they should be in order.
+#' @param ... One or more `SpatRaster`, `SpatRasterCollection`, `SpatVector`, `SpatVectorProxy` or `SpatExtent` objects (whose combined bounding box extent will be returned); or the following _named_ numeric arguments: `xmin`/`ymax`/`xmax`/`ymin`. If these four limit arguments are not named they should be in the stated order.
+#' 
 #' @return a _list_ object describing a GeoJSON bounding rectangular polygon suitable for use as `regions` argument to `gd_download()` or `gd_search()`
+#' 
 #' @export
 #' @examples
 #' gd_bbox(
@@ -27,15 +29,44 @@ gd_bbox <- function(...) {
              x[["xmax"]], x[["ymin"]]),
            ncol = 2, byrow = TRUE)
   }
-  # TODO: bbox around all args for multiple SpatExtent?
-  if (inherits(.args[[1]], 'SpatExtent')) {
+  
+  # if ... has an ext() method we will use it
+  tri <- vapply(.args, inherits, logical(1), c(
+    "SpatRaster",
+    "SpatRasterCollection",
+    "SpatVector",
+    "SpatVectorProxy",
+    "SpatExtent"
+  ))
+  
+  if (all(tri)) {
+    
     stopifnot(requireNamespace("terra", quietly = TRUE))
-    m <- do.call('cbind', .args[[1]]@ptr$as.points())
+    
+    
+    # extend extent for each terra object in ...
+    if (!inherits(.args[[1]], 'SpatExtent')) {
+      # convert non-terra to terra
+      out <- .cast_spatial_object(.args[[1]], extent = TRUE)
+    } else {
+      out <- .args[[1]]
+    }
+    
+    if (length(.args) > 1) {
+      for (a in .args[-1]) {
+        if (!inherits(a, 'SpatExtent')) {
+          a <- terra::ext(a)
+        }
+        out <- terra::union(out, a)
+      }
+    }
+    m <- terra::crds(terra::as.points(out))
+      
   } else if (all(.gdal_projwin %in% names(.args))) {
     m <- .mbbox(.args)
   } else {
     if (!length(.args) == 4) {
-      stop('Expecting total of 4 bounding box arguments, If arguments are unnamed they should be in the following order: "xmin", "ymax", "xmax", "ymin".', call. = FALSE)
+      stop('Expecting total of 4 numeric bounding box arguments.\nIf arguments are unnamed they should be in the following order: "xmin", "ymax", "xmax", "ymin".', call. = FALSE)
     }
     names(.args) <- .gdal_projwin
     m <- .mbbox(.args)
@@ -59,13 +90,20 @@ gd_bbox <- function(...) {
 #' @return list representing a GeoJSON extent
 #' @importFrom  methods as
 #' @export
-#' @examplesIf requireNamespace("terra")
+#' @examplesIf !inherits(requireNamespace("terra", quietly=TRUE), 'try-error')
 #' @examples
-#' b <- terra::vect('POLYGON((-121.355 37.56,-121.355 37.555,
-#'                     -121.35 37.555,-121.35 37.56,
-#'                     -121.355 37.56))',
-#'           crs = "OGC:CRS84")
+#' \donttest{
+#' library(terra)
+#' 
+#' b <- terra::vect('POLYGON((-121.355 37.560, 
+#'                            -121.355 37.555,
+#'                            -121.350 37.555, 
+#'                            -121.350 37.560,
+#'                            -121.355 37.560))',
+#'                  crs = "OGC:CRS84")
+#'
 #' gd_region(b)
+#' }
 gd_region <- function(x) {
 
   if (is.list(x) &&
@@ -80,40 +118,14 @@ gd_region <- function(x) {
          See `gd_bbox()` for a simpler region interface that takes numeric values (xmin/xmax/ymin/ymax) directly.", .call = FALSE)
   }
 
-  # wkt string
-  if (is.character(x)) {
-    x <- terra::vect(x, crs = "OGC:CRS84")
-  }
-
-  # raster/sp support
-  if (inherits(x, 'Spatial')) {
-    if (requireNamespace('raster', quietly = TRUE)) {
-      x <- terra::vect(as(x, 'Spatial'))
-    }
-  }
-
-  if (inherits(x, 'RasterLayer') || inherits(x, 'RasterStack')) {
-    if (requireNamespace('raster', quietly = TRUE)) {
-      x <- terra::rast(x)
-    }
-  }
-
-  # sf objects
-  if (inherits(x, c('sf', 'sfc'))) {
-    x <- terra::vect(x)
-  }
-
+  # convert non-terra to terra
+  x <- .cast_spatial_object(x)
+  
   # terra
   if (inherits(x, c('SpatVector',
                     'SpatRaster',
                     'SpatVectorCollection',
                     'SpatRasterCollection'))) {
-    if (!inherits(x, 'SpatVector')) {
-      x <- terra::as.polygons(x, extent = TRUE)
-      # project everything. will fail if CRS in x not defined.
-      x <- terra::project(x, "OGC:CRS84")
-
-    }
     return(.gd_geojson(x))
   }
   gd_bbox(x)
@@ -132,4 +144,62 @@ gd_region <- function(x) {
   list(type = "Polygon", coordinates = list(lapply(apply(p, 1, function(x) {
     list(as.numeric(x))
   }), .subset2, 1)))
+}
+
+#' Cast Spatial Object to SpatVector or SpatRaster
+#' 
+#' This internal function allows for consistent interfaces for non-terra Spatial objects by coercion to the terra native objects `SpatVector` or `SpatRaster`.
+#' 
+#' @param x A WKT string, Spatial*, Raster*, or sf* object
+#' @param extent Return only SpatExtent of result? Default: `FALSE`
+#' @details WKT string coordinates should use the longitude latitude WGS84 decimal degrees (`"OGC:CRS84"` spatial reference system).
+#' @noRd
+.cast_spatial_object <- function(x, extent = FALSE) {
+  
+  # wkt string
+  if (is.character(x)) {
+    x <- terra::vect(x, crs = "OGC:CRS84")
+  }
+  
+  # raster/sp support
+  if (inherits(x, 'Spatial')) {
+    if (requireNamespace('raster', quietly = TRUE)) {
+      x <- terra::vect(as(x, 'Spatial'))
+    }
+  }
+  
+  if (inherits(x, c('RasterLayer', 'RasterStack'))) {
+    if (requireNamespace('raster', quietly = TRUE)) {
+      x <- terra::rast(x)
+    }
+  }
+  
+  # sf and sfc objects
+  if (inherits(x, c('sf', 'sfc'))) {
+    x <- terra::vect(x)
+  }
+  
+  # convert to simple geometries if we only want extent
+  if (extent && !inherits(x, 'SpatVector')) {
+    x <- terra::as.polygons(x, extent = TRUE)
+  }
+  
+  # project what we can to OGC:CRS84
+  if (inherits(x, 'SpatVector')) {
+    # will fail if CRS in x not defined
+    x <- try(terra::project(x, "OGC:CRS84"), silent = TRUE)
+    if (inherits(x, 'try-error')) {
+      stop(x[1], call. = FALSE)
+    }
+  }
+  
+  # raster Extent, sf bbox, extent=TRUE
+  # assume these are already in correct CRS
+  if (inherits(x, c('Extent', 'bbox')) || 
+      (extent && !inherits(x, 'SpatExtent'))) {
+    x <- terra::ext(x)
+  }
+  
+  # return object (possibly unchanged)
+  x
 }